refactor: consolida autenticação em auth-service
- Move método signIn de authService.ts para auth-service.ts
- Remove referências a localStorage (token agora em cookie httpOnly)
- Remove interface LoginData não utilizada, usa SigninFormType do zod
- Adiciona função signOut para limpar cookie
- Remove arquivo authService.ts duplicado
- Atualiza LogoutButton para usar server action signOut
