+++ src/features/auth/index.ts
@@ -1,3 +1,3 @@
 export * from './components';
-export { AuthService } from './services/auth-service';
+export { signIn, signOut } from './services/auth-service';
 export { LogoutButton } from './components/logout-button';
