+++ src/features/auth/services/auth-service.ts
@@ -1,62 +1,47 @@
-const API_BASE_URL = 'http://localhost:8080';
+'use server';
 
-// Interface para dados de login
-interface LoginData {
-	matricula: string;
-	senha: string;
+import { apiService } from '@/features/core/services/apiService';
+import { SigninFormType } from '../schemas/signinForm.schema';
+import { cookies } from 'next/headers';
+import SigninResponse from '../types/signinResponse';
+import { jwtDecode } from 'jwt-decode';
+import { encrypt } from '@/lib/auth';
+import { DecodedToken } from '@/features/core/types/decodedJwt';
+
+export async function signIn(payload: SigninFormType) {
+	try {
+		const data = await apiService<SigninResponse>('/auth/login', {
+			method: 'POST',
+			body: payload,
+		});
+
+		const cookieStore = await cookies();
+
+		const decoded = jwtDecode(data.token) as DecodedToken;
+		const expiresInMs = (decoded.exp as number) * 1000;
+		const token = await encrypt(
+			{ ...decoded, token: data.token },
+			expiresInMs
+		);
+
+		cookieStore.set('token', token, {
+			path: '/',
+			expires: new Date(expiresInMs),
+			secure: false,
+			httpOnly: true,
+		});
+
+		return {
+			name: decoded.sub,
+			role: decoded.role,
+		};
+	} catch (error) {
+		console.error('Erro ao realizar login: ', error);
+		throw error;
+	}
 }
 
-// Interface para resposta de login
-interface LoginResponse {
-	token: string;
-}
-
-export class AuthService {
-	static async login(matricula: string, senha: string): Promise<string> {
-		try {
-			const response = await fetch(`${API_BASE_URL}/auth/login`, {
-				method: 'POST',
-				headers: {
-					'Content-Type': 'application/json',
-				},
-				body: JSON.stringify({ matricula, senha }),
-			});
-
-			if (!response.ok) {
-				if (response.status === 401) {
-					throw new Error('Matrícula ou senha incorretos');
-				}
-				throw new Error(`Erro no login: ${response.status}`);
-			}
-
-			const data: LoginResponse = await response.json();
-			return data.token;
-		} catch (error) {
-			console.error('Erro ao fazer login:', error);
-			throw error;
-		}
-	}
-
-	static saveToken(token: string): void {
-		if (typeof window !== 'undefined') {
-			localStorage.setItem('authToken', token);
-		}
-	}
-
-	static getToken(): string | null {
-		if (typeof window !== 'undefined') {
-			return localStorage.getItem('authToken');
-		}
-		return null;
-	}
-
-	static removeToken(): void {
-		if (typeof window !== 'undefined') {
-			localStorage.removeItem('authToken');
-		}
-	}
-
-	static isAuthenticated(): boolean {
-		return this.getToken() !== null;
-	}
+export async function signOut() {
+	const cookieStore = await cookies();
+	cookieStore.delete('token');
 }
